+++ kakao-gift/craco.config.js
@@ -7,6 +7,7 @@ module.exports = {
       '@styles': path.resolve(__dirname, 'src/styles'),
       '@components': path.resolve(__dirname, 'src/components'),
       '@utils': path.resolve(__dirname, 'src/utils'),
+      '@stories': path.resolve(__dirname, 'src/stories'),
     },
   },
 }
