+++ kakao-gift/src/stories/GoodsItem/GoodsItem.stories.tsx
