+++ kakao-gift/src/stories/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Button } from './Button'
+import { Button } from '../Button/Button'
 import './header.css'
 
 type User = {
