+++ kakao-gift/src/stories/Image/Image.stories.tsx
