+++ kakao-gift/src/stories/Image/Image.tsx
