+++ kakao-gift/src/stories/Input/Input.tsx
