+++ kakao-gift/tsconfig.paths.json
@@ -5,7 +5,8 @@
       "@src/*": ["src/*"],
       "@styles/*": ["src/styles/*"],
       "@components/*": ["src/components/*"],
-      "@utils/*": ["src/utils/*"]
+      "@utils/*": ["src/utils/*"],
+      "@stories/*": ["src/stories/*"]
     }
   }
 }
